A lot of time was being spent in String object creation
Turns out we don't need to create new String for every line in the file.

We are creating byte array for the bytes of the location name, we can
compute the hash of these bytes to see if an entry exists for this
location in the table or not. And the table entry itself doesn't need to
store the String object itself, it can simply store a copy of the
location name byte array.