Avoid copying the location name bytes to the byte array.
The copygint to the byte array in readFile was unnecessary, it would
only cause extra writes in memory.

Later on in processLine when we have to compare the bytes again, it
doesn't matter whether we are reading them from the byte array, or from
the file, it is coming from L1. So we might as well avoid the byte
array.

It sure makes the logic for collision check ugly because now we are
having to pull the bytes from memory and compare them in three different
places.